+++ app/Http/Controllers/ViewInventoryController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use App\Genre;
 use App\InventoryItem;
 
 class ViewInventoryController extends Controller
@@ -10,6 +10,7 @@ class ViewInventoryController extends Controller
     public function index()
     {
         $inventoryItems = InventoryItem::allJoined();
-        return view('view-inventory', compact('inventoryItems'));
+        $genres = Genre::all();
+        return view('view-inventory', compact('inventoryItems', 'genres'));
     }
 }