+++ database/migrations/2018_09_21_082556_create_genre_inventory_item_table.php
@@ -1,0 +1,42 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateGenreInventoryItemTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('genre_inventory_item', function (Blueprint $table) {
+            $table->increments('id');
+            $table->integer('inventory_item_id')->unsigned();
+            $table->integer('genre_id')->unsigned();
+
+            $table->foreign('inventory_item_id')
+                ->references('id')
+                ->on('inventory_items')
+                ->onDelete('cascade');
+
+            $table->foreign('genre_id')
+                ->references('id')
+                ->on('genres')
+                ->onDelete('cascade');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('genre_inventory_item');
+    }
+}