+++ database/seeds/DatabaseSeeder.php
@@ -18,8 +18,12 @@ class DatabaseSeeder extends Seeder
         $this->command->info('User roles table seeded.');
 
         if (App::environment() === 'local') {
+            $this->call(GenresSeeder::class);
+            $this->command->info('Genres table seeded. (local environment)');
             $this->call(InventoryItemsSeeder::class);
             $this->command->info('Inventory items table seeded. (local environment)');
+            $this->call(GenreInventoryItemSeeder::class);
+            $this->command->info('Inventory item genre pivot table seeded. (local environment)');
             $this->call(UsersSeeder::class);
             $this->command->info('Users table seeded. (local environment)');
         }