+++ database/seeds/InventoryItemsSeeder.php
@@ -16,10 +16,26 @@ class InventoryItemsSeeder extends Seeder
     {
         Schema::disableForeignKeyConstraints();
         DB::table('inventory_items')->truncate();
-        InventoryItem::create(['name' => 'Dixit', 'status_id' => 1]);
-        InventoryItem::create(['name' => 'Shadow Hunters', 'status_id' => 1]);
-        InventoryItem::create(['name' => 'Puissance 4 géant', 'status_id' => 2]);
-        InventoryItem::create(['name' => 'Wii U', 'status_id' => 1]);
+        InventoryItem::create([
+            'name' => 'Dixit',
+            'status_id' => 1,
+            'duration_min' => 10,
+            'duration_max' => 60,
+            'players_min' => 3,
+            'players_max' => 12]);
+        InventoryItem::create([
+            'name' => 'Shadow Hunters',
+            'status_id' => 1,
+            'duration_min' => 20,
+            'duration_max' => 60,
+            'players_min' => 4,
+            'players_max' => 9]);
+        InventoryItem::create([
+            'name' => 'Puissance 4 géant',
+            'status_id' => 2]);
+        InventoryItem::create([
+            'name' => 'Wii U',
+            'status_id' => 1]);
         Schema::enableForeignKeyConstraints();
     }
 }